src/ondalear/backend/requirements/common.txt
Babel==2.7.0
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
django-constrainedfilefield==3.2.0
django-extensions==2.1.9
django-filter==2.2.0
django-phonenumber-field==3.0.1
django-rest-auth==0.9.5
django-timezone-field==3.0
djangorestframework==3.9.4
djangorestframework-filters==1.0.0.dev0
djangorestframework-jwt==1.11.0
ftfy==5.5.1
httpie==1.0.2
idna==2.8
inflection==0.3.1
jsonfield==2.0.2
phonenumbers==8.10.14
pip==19.2.1
psycopg2==2.8.3
Pygments==2.4.2
PyJWT==1.7.1
python-magic==0.4.15
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3

